JCR-5207: Deprecate TransientFileFactory (also add test hook)

+++ jackrabbit-jcr-commons/src/main/java/org/apache/jackrabbit/util/TransientFileFactory.java
@@ -22,7 +22,6 @@ import java.lang.ref.PhantomReference;
 import java.lang.ref.ReferenceQueue;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Iterator;
 import java.util.Collections;
 
 /**
@@ -45,14 +44,14 @@ public class TransientFileFactory {
      * Queue where <code>MoribundFileReference</code> instances will be enqueued
      * once the associated target <code>File</code> objects have been gc'ed.
      */
-    private final ReferenceQueue<File> phantomRefQueue = new ReferenceQueue<File>();
+    private final ReferenceQueue<File> phantomRefQueue = new ReferenceQueue<>();
 
     /**
      * Collection of <code>MoribundFileReference</code> instances currently
      * being tracked.
      */
     private final Collection<MoribundFileReference> trackedRefs =
-        Collections.synchronizedList(new ArrayList<MoribundFileReference>());
+        Collections.synchronizedList(new ArrayList<>());
 
     /**
      * The reaper thread responsible for removing files awaiting deletion
@@ -87,11 +86,7 @@ public class TransientFileFactory {
         reaper.start();
         // register shutdownhook for final cleaning up
         try {
-            shutdownHook = new Thread() {
-                public void run() {
-                    doShutdown();
-                }
-            };
+            shutdownHook = new Thread(this::doShutdown);
             Runtime.getRuntime().addShutdownHook(shutdownHook);
         } catch (IllegalStateException e) {
             // can't register shutdownhook because
@@ -134,7 +129,7 @@ public class TransientFileFactory {
      * the webapp classloader. This must be called after all repositories had
      * been stopped, so use with great care!
      * <p>
-     * See http://issues.apache.org/jira/browse/JCR-1636 for details.
+     * See <a href="http://issues.apache.org/jira/browse/JCR-1636">...</a> for details.
      */
     public static void shutdown() {
         getInstance().doShutdown();
@@ -150,8 +145,8 @@ public class TransientFileFactory {
         // to avoid ConcurrentModificationException (JCR-549)
         // @see java.lang.util.Collections.synchronizedList(java.util.List)
         synchronized(trackedRefs) {
-            for (Iterator<MoribundFileReference> it = trackedRefs.iterator(); it.hasNext();) {
-                it.next().delete();
+            for (MoribundFileReference trackedRef : trackedRefs) {
+                trackedRef.delete();
             }
 
         }
@@ -198,10 +193,13 @@ public class TransientFileFactory {
                     // silently ignore...
                     continue;
                 }
+
                 // delete target
-                fileRef.delete();
-                fileRef.clear();
-                trackedRefs.remove(fileRef);
+                if (fileRef != null) {
+                    fileRef.delete();
+                    fileRef.clear();
+                    trackedRefs.remove(fileRef);
+                }
             }
         }
